+++ public/index.html
@@ -18,11 +18,7 @@
 </header>
 
 <main class="main">
-  <!-- Меню -->
 
-  <!-- Сортировка -->
-
-  <!-- Контент -->
 </main>
