+++ src/components/container.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils/render.js";
+import AbstractComponent from "./abstract-component.js";
 
 
 const createFilmsContainerTemplate = () => {
@@ -32,24 +32,9 @@ const createFilmsContainerTemplate = () => {
   );
 };
 
-export default class FilmsContainer {
-  constructor() {
-    this._element = null;
-  }
 
+export default class FilmsContainer extends AbstractComponent {
   getTemplate() {
     return createFilmsContainerTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
