+++ src/components/footer-statistic.js
@@ -1,31 +1,18 @@
-import {createElement} from "../utils/render.js";
-
+import AbstractComponent from "./abstract-component.js";
 
 const createMovieStaisticTemplate = (movieStaistic) => {
   return (`<p>${movieStaistic.length} movies inside</p>`);
 };
 
 
-export default class MovieStaistic {
+export default class MovieStaistic extends AbstractComponent {
   constructor(movieStaistic) {
+    super();
     this._movieStaistic = movieStaistic;
 
-    this._element = null;
   }
 
   getTemplate() {
     return createMovieStaisticTemplate(this._movieStaistic);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
