+++ src/utils/removeChild.js
@@ -1,0 +1,3 @@
+export const removeElement = (container, element) => {
+  container.removeChild(element.getElement());
+};
